Update README.md
Made a clarification in the README about the abcde conf file line 66 permission issue. If the system is not created with the arm user in mind but rather added on (as show in the beginning of the walk through) the final destination output directory might not have the appropriate permissions. This edit is to clarify that this location is set appropriately.